Split stages into separate files
